requirements.txt
coverage==7.11.0
iniconfig==2.3.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
